Break out utils tests
